use common smrtflow version code from smrt-common-models
