+++ smrt-analysis/src/main/scala/com/pacbio/secondary/analysis/converters/ReferenceInfoConverter.scala
@@ -11,7 +11,7 @@ import org.joda.time.DateTime
 import collection.JavaConverters._
 import collection.JavaConversions._
 import com.pacbio.common.models.{Constants => CommonConstants}
-import com.pacbio.secondary.analysis.constants.{FileTypes, GlobalConstants}
+import com.pacbio.secondary.analysis.constants.FileTypes
 import com.pacbio.secondary.analysis.datasets._
 import com.pacbio.secondary.analysis.legacy.{ReferenceContig, ReferenceEntry, ReferenceEntryIO, ReferenceInfoType}
 import com.pacificbiosciences.pacbiobasedatamodel.IndexedDataType.FileIndices
@@ -209,7 +209,7 @@ object ReferenceInfoConverter extends LazyLogging {
     val r = datasetIO.dataset
     val erUUID = UUID.randomUUID()
 
-    val author = s"pbscala ${GlobalConstants.PB_SCALA_VERSION} reference_info_dataset_" + REF_INFO_TO_DS_VERSION
+    val author = s"pbscala ${CommonConstants.SMRTFLOW_VERSION} reference_info_dataset_" + REF_INFO_TO_DS_VERSION
 
     val root =
       <pbds:ReferenceSet
