+++ smrt-server-base/src/main/scala/com/pacbio/common/actors/StatusServiceActor.scala
@@ -4,6 +4,7 @@ import java.util.{Properties, UUID}
 
 import akka.actor.{ActorRef, Props}
 import com.pacbio.common.dependency.Singleton
+import com.pacbio.common.models.Constants
 import com.pacbio.common.time.{Clock, ClockProvider}
 import org.joda.time.{Duration => JodaDuration, Instant => JodaInstant}
 import scala.collection.JavaConverters._
@@ -61,24 +62,7 @@ sealed trait BaseStatusServiceActorProvider {
 
   val uuid: Singleton[UUID] = Singleton(UUID.randomUUID())
 
-  val buildVersion: Singleton[String] = Singleton(() => {
-    val files = getClass().getClassLoader().getResources("version.properties")
-    if (files.hasMoreElements) {
-      val in = files.nextElement().openStream()
-      try {
-        val prop = new Properties
-        prop.load(in)
-        prop.getProperty("version").replace("SNAPSHOT", "") + prop.getProperty("sha1").substring(0, 7)
-      }
-      finally {
-        in.close()
-      }
-    }
-    else {
-      "unknown version"
-    }
-  }
-  )
+  val buildVersion: Singleton[String] = Singleton(() => Constants.SMRTFLOW_VERSION)
 }
